requirements.txt
# AI Lead Generation System - Python Dependencies
anthropic>=0.39.0
openai>=1.50.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
tenacity>=8.2.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
